requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
flask==3.1.0
Flask-Login==0.6.3
flask-wtf==1.2.2
flask-restx==1.3.0
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
python-dotenv==1.1.0
sqlalchemy==2.0.40
typing-extensions==4.13.2
typing-inspection==0.4.0
werkzeug==3.1.3
wtforms==3.2.1
zipp==3.21.0
